requirements.txt
gunicorn>=20.0.4
Flask>=2.3.3
requests>=2.23.0
Flask-Bootstrap>=3.3.7.1
Flask-WTF>=1.2.1
Pillow>=9.5.0 --only-binary=all
lxml>=4.9.4
tensorflow-cpu>=2.13.0
keras>=2.13.1
opencv-python-headless>=4.8.1.78
numpy>=1.24.3
matplotlib>=3.7.2
scikit-image>=0.21.0
